requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
attrs==20.2.0
Automat==0.6.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cloud-init==19.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
DateTime==4.3
decorator==4.1.2
distro-info===0.18ubuntu0.18.04.1
Django==3.0.2
gunicorn==20.0.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
importlib-metadata==2.0.0
incremental==16.10.1
iniconfig==1.0.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
minidump==0.0.6
minikerberos==0.0.10
netifaces==0.10.4
numpy==1.13.3
oauthlib==2.0.6
olefile==0.45.1
packaging==20.4
PAM==0.4.2
Pillow==5.1.0
pluggy==0.13.1
prettytable==0.7.2
py==1.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pypykatz==0.0.8
pyserial==3.4
pytest==6.1.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.4
pyxdg==0.25
PyYAML==3.12
requests==2.25.1
requests-unixsocket==0.1.5
scapy==2.4.3
scipy==0.19.1
SecretStorage==2.3.1
service-identity==16.0.0
shodan==1.24.0
six==1.15.0
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
toml==0.10.1
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
virtualenv==16.7.9
XlsxWriter==1.3.7
zipp==3.3.0
zope.interface==5.2.0
